Move all shared logic from review.py into the clang_tidy_review library
